+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Fixed:
 
 Added:
   - `mappersmith`: Make Response accept a generic type that specifies the form of the data returned #265
+  - `mappersmith`: Add the possibility to update the default encoding function for query params #296
   - `mappersmith/test`: Add `responseFactory` and `requestFactory` helpers to `mappersmith/test` #265
   - `mappersmith/test`: If `body` is not provided to a mock, it will match on any body. #229 #152